chore: move database from /internal
