+++ products/postgres/repository_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
-	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/database"
+	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/internal/database"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/products"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/products/postgres"
 	"gorm.io/gorm"
